fix!: Remove unnecessary fee payment to bootloader address (#340)
## What ❔
This PR removes the intermediate step of transferring fees to the
bootloader formal address. Now, fees are payed directly to the coinbase.
In addition, this PR removes some unnecessary logic around fee payment,
which is some legacy from AA support and it moves the actual payment of
the fee to the operator to the end of the transaction.
<!-- What are the changes this PR brings about? -->
<!-- Example: This PR adds a PR template to the repo. -->
<!-- (For bigger PRs adding more context is appreciated) -->

## Why ❔
It simplifies the code and also fixes an issue related to pubdata
accounting for the change in bootloader balance.
<!-- Why are these changes done? What goal do they contribute to? What
are the principles behind them? -->
<!-- The `Why` has to be clear to non-Matter Labs entities running their
own ZK Chain -->
<!-- Example: PR templates ensure PR reviewers, observers, and future
iterators are in context about the evolution of repos. -->

## Is this a breaking change?
- [X] Yes
- [ ] No

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [X] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [X] Tests for the changes have been added / updated.
- [X] Documentation comments have been added / updated.
- [X] Code has been formatted.

---------

Co-authored-by: Vladislav Volosnikov <Volosnikov.apmath@gmail.com>