fix: Fix overflow issue in modexp (#385)
## What ❔
The code assumes that adding carries never overflows existing values
from c, but c could be the maximum value. Instead of asserting no
overflow, we need to propagate the overflow to the next digit
<!-- What are the changes this PR brings about? -->
<!-- Example: This PR adds a PR template to the repo. -->
<!-- (For bigger PRs adding more context is appreciated) -->

## Why ❔
These fix a correctness bug in the delegated bigint multiplication
routine. When the schoolbook multiply finished iterating over the limbs,
the carry left in carry_scratch sometimes needed to ripple into a new
high limb. Previously that last propagation step was effectively ignored
once we exited the inner loops. As a result, the produced product could
miss an entire high word, leading to mismatches.

<!-- Why are these changes done? What goal do they contribute to? What
are the principles behind them? -->
<!-- The `Why` has to be clear to non-Matter Labs entities running their
own ZK Chain -->
<!-- Example: PR templates ensure PR reviewers, observers, and future
iterators are in context about the evolution of repos. -->

## Is this a breaking change?
- [ ] Yes
- [X] No

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [X] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [X] Tests for the changes have been added / updated.
- [ ] Documentation comments have been added / updated.
- [X] Code has been formatted.

---------

Co-authored-by: Vladislav Volosnikov <Volosnikov.apmath@gmail.com>