fix!: do not charge for native resources when out of gas (#422)
## What ❔
This PR changes how we charge for resources: now we don't charge for
native if the ergs charge leads to an Out of Ergs error.
<!-- What are the changes this PR brings about? -->
<!-- Example: This PR adds a PR template to the repo. -->
<!-- (For bigger PRs adding more context is appreciated) -->

## Why ❔
The `charge` function is designed to charge for both ergs and native at
the same time. This function is always called as pre-charging (before
the computation being charged is executed). As such, it's safe to not
charge for native when OOG, as the computation is not performed anyways.
That way, we can keep evm equivalence in the case where both ergs and
native would run out for a single `charge` call.
<!-- Why are these changes done? What goal do they contribute to? What
are the principles behind them? -->
<!-- The `Why` has to be clear to non-Matter Labs entities running their
own ZK Chain -->
<!-- Example: PR templates ensure PR reviewers, observers, and future
iterators are in context about the evolution of repos. -->

## Is this a breaking change?
- [x] Yes
- [ ] No

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [ ] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [ ] Tests for the changes have been added / updated.
- [ ] Documentation comments have been added / updated.
- [ ] Code has been formatted.